% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_deseq_table.R
\name{annotate_deseq_table}
\alias{annotate_deseq_table}
\title{Annotate and Transform DESeq2 Results Table}
\usage{
annotate_deseq_table(
  deseq_results_list,
  params,
  filter_results = FALSE,
  biosets_filter = FALSE
)
}
\arguments{
\item{deseq_results_list}{A list of data frames, each containing DESeq2 results.}

\item{params}{List containing parameters such as platform, feature_id, species_data, etc.}

\item{filter_results}{Logical; if TRUE, apply additional filters to the table.}

\item{biosets_filter}{Logical; if TRUE, apply filters specifically for 'biosets'.}
}
\value{
An annotated and transformed data frame of DESeq2 results.
}
\description{
This function takes a list of DESeq2 results and annotates them with gene symbols
and descriptions. It performs transformation of log2 fold changes into linear
fold changes and applies optional filters.
}
