% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_DESeq_results.R
\name{apply_quant_spike_filters}
\alias{apply_quant_spike_filters}
\title{Apply R-ODAF Quantile and Spike Filters to DESeq2 counts}
\usage{
apply_quant_spike_filters(DECounts, exp_metadata_subset, SampPerGroup, design)
}
\arguments{
\item{DECounts}{Matrix or data frame with DEGs and their corresponding count data.}

\item{exp_metadata_subset}{Data frame containing the experimental metadata subset for the current comparison.}

\item{SampPerGroup}{Named vector with the count of samples per group.}

\item{design}{The design variable that separates groups in the metadata subset.}
}
\value{
A matrix with logical values indicating whether each gene passed the quantile
and spike filters, respectively ('1' indicates that the gene passed the filter).
}
\description{
This function applies post-processing filters to differentially expressed genes (DEGs).
It checks whether DEGs pass the median vs third quantile check and spike condition.
}
