% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_DESeq_results.R
\name{apply_relevance_filter}
\alias{apply_relevance_filter}
\title{Apply Relevance Filter to Genes Based on Counts}
\usage{
apply_relevance_filter(
  CPMdds,
  exp_metadata_subset,
  SampPerGroup,
  MinCount,
  design,
  threshold_proportion = 0.75
)
}
\arguments{
\item{CPMdds}{Matrix of normalized counts as CPM (Counts Per Million).}

\item{exp_metadata_subset}{Data frame containing the experimental metadata subset for the current comparison.}

\item{SampPerGroup}{Named vector with the count of samples per condition/group.}

\item{MinCount}{The minimum count threshold to consider a gene as relevant.}

\item{threshold_proportion}{The proportion of samples within a group that must exceed the count threshold.}
}
\value{
Logical vector indicating whether each gene has passed the relevance filter.
}
\description{
This function filters genes based on a minimum count threshold. For a gene to pass the filter,
a specified proportion of its counts must be above the threshold in at least one of the conditions.
}
