% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data_and_metadata.R
\name{check_data}
\alias{check_data}
\title{Perform Sanity Checks on Data}
\usage{
check_data(sd, exp_metadata, exp_contrasts)
}
\arguments{
\item{sd}{A data frame of count data to check.}

\item{exp_metadata}{A data frame of the experimental metadata to check against count data.}

\item{exp_contrasts}{A data frame of contrasts to validate.}
}
\value{
None; this function is called for its side effects.
}
\description{
Runs a series of sanity checks on count data and corresponding metadata
to ensure that they are consistent with each other for further analysis.
}
