% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_up_paths.R
\name{create_dirs_recursively}
\alias{create_dirs_recursively}
\title{Create directories from paths recursively.}
\usage{
create_dirs_recursively(paths)
}
\description{
Create directories from paths recursively.
}
