% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_report.R
\name{create_random_string}
\alias{create_random_string}
\title{Create a random string.}
\usage{
create_random_string(len = 20, digits = 0:9, letters_set = c(LETTERS, letters))
}
\arguments{
\item{digits}{A vector of digits to include in the random string.
Default is 0:9.}

\item{letters_set}{A vector of letters to include in the random string.}
}
\value{
A random string of length 10.
}
\description{
This function generates a random string by sampling characters from the
combination of uppercase letters and digits.
}
