% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data_and_metadata.R
\name{filter_data}
\alias{filter_data}
\title{Filter Count Data Based on Minimum Counts}
\usage{
filter_data(count_data, exp_metadata, threshold)
}
\arguments{
\item{count_data}{A data frame containing count data.}

\item{exp_metadata}{A data frame containing the corresponding experimental metadata.}

\item{threshold}{A numeric threshold for minimum allowed count sum across samples.}
}
\value{
A filtered count data matrix or data frame.
}
\description{
Removes samples from count data that have total read counts below a specified threshold.
NA values within the count data are replaced with zeros.
}
