% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data_and_metadata.R
\name{filter_metadata}
\alias{filter_metadata}
\title{Filter experimental metadata based on inclusion/exclusion criteria}
\usage{
filter_metadata(exp_metadata, params, exp_contrasts)
}
\arguments{
\item{exp_metadata}{A data frame containing the experiment metadata.}

\item{params}{The list of parameters describing the study.}

\item{design}{The name of the column used for the experimental design grouping.}
}
\value{
A filtered metadata data frame.
}
\description{
This function processes the experimental metadata for an
experiment by excluding or including certain samples or groups based
on user-defined criteria.
}
