% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data_and_metadata.R
\name{format_and_sort_metadata}
\alias{format_and_sort_metadata}
\title{Format and sort experiment metadata for downstream analysis}
\usage{
format_and_sort_metadata(exp_metadata, design, sortcol)
}
\arguments{
\item{exp_metadata}{A data frame of the experiment metadata.}

\item{design}{The name of the column used for the experimental design grouping.}

\item{sortcol}{An optional column used for sorting the \code{design} grouping.}

\item{intgroup}{A character vector of column names to be converted to factors for grouping.}
}
\value{
The formatted and sorted metadata data frame.
}
\description{
This function formats and sorts the experimental metadata by converting
grouping variables (\code{intgroup}) to factors and sorting by another
column (\code{sortcol}) if specified.
}
