% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_analysis_id.R
\name{get_analysis_id}
\alias{get_analysis_id}
\title{Generate an Analysis ID}
\usage{
get_analysis_id(params)
}
\arguments{
\item{params}{A list of parameters from which the ID is constructed.}
}
\value{
A character string representing the unique analysis ID.
}
\description{
This function generates a unique ID for the analysis based on the current time and the project title.
If a \code{deseq_filter} is provided, it is appended to the ID string.
}
