% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_clean_names.R
\name{get_clean_names}
\alias{get_clean_names}
\title{Obtain Clean Names for Visualization}
\usage{
get_clean_names(res, vector_of_names, feature_id, species_gene_symbol)
}
\arguments{
\item{res}{Data frame containing the annotation information.}

\item{vector_of_names}{Character vector of feature identifiers to map to gene symbols.}

\item{feature_id}{The feature identifier column name in res used for mapping to vector_of_names.}

\item{species_gene_symbol}{The gene symbol column name in res used for mapping to gene names.}
}
\value{
A character vector of clean names suitable for heatmaps or other visualizations.
}
\description{
This function joins input names with annotations to create a clean set of identifiers, such as gene symbols,
for use in visualizations like heatmaps. It handles cases where multiple genes are mapped from a single feature.
}
