% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn_deseq_model.R
\name{get_design}
\alias{get_design}
\title{Generate Design Formula for DESeq2 Analysis}
\usage{
get_design(design, params)
}
\arguments{
\item{design}{Character vector specifying design factor(s).}

\item{params}{A list of parameters containing formula_override.}
}
\value{
A design formula for DESeq2.
}
\description{
This function generates a design formula for use in DESeq2 analysis.
If a formula override is provided in the params, that formula is used instead.
}
