% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_facets.R
\name{get_facets}
\alias{get_facets}
\title{Get Unique Facets from Metadata}
\usage{
get_facets(
  metadata = exp_metadata,
  exclude = params$exclude_groups,
  report_facet = params$reports_facet,
  skip_extra = "DMSO"
)
}
\arguments{
\item{metadata}{A data frame containing the experimental metadata.}

\item{exclude}{A character vector of group names to be excluded from the facets.}

\item{report_facet}{A character string specifying the column name in \code{metadata}
from which to extract the facets.}

\item{skip_extra}{A character string specifying an additional group name to be
excluded from the facets.}
}
\value{
A character vector of unique facet values after exclusion.
}
\description{
This function extracts unique values from a specified column in the metadata,
excluding any groups specified by the user and an additional skip value.
}
