% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata.R
\name{get_metadata}
\alias{get_metadata}
\title{Read Experiment Metadata}
\usage{
get_metadata(file_path = NULL, paths)
}
\arguments{
\item{file_path}{A string indicating the file path of the metadata file to read.}

\item{paths}{A list of paths used throughout the analysis.}
}
\value{
A data frame containing the experiment metadata with an additional
column, 'original_names', which stores the original row names.
}
\description{
Reads the experimental metadata from a specified file and adds a
new column with the original row names.
}
