% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_params.R
\name{get_params}
\alias{get_params}
\title{Retrieve and Set Up Parameters for a Given Context}
\usage{
get_params(context = "analysis")
}
\arguments{
\item{context}{Character string specifying the context for parameter setup; accepts 'analysis' or 'QC'.}
}
\value{
A list of parameters set up for the specified context.
}
\description{
This function reads configuration settings from a YAML file and sets up the parameters for the
specified context, either 'analysis' or 'QC'. It prepares the parameters list by combining common
settings with context-specific settings and replaces any NULLs with NA. It also determines the
project directory and loads species-specific data.
}
