% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prefix.R
\name{get_prefix}
\alias{get_prefix}
\title{Generate a Filename Prefix}
\usage{
get_prefix(prefix_pars, prefix_facet)
}
\arguments{
\item{prefix_pars}{A list containing parameters used to construct the prefix,
including 'platform', 'project_title', and optionally 'reports_facet'
and 'reports_filter'.}

\item{prefix_facet}{An optional character string representing the facet to include
in the prefix. If 'reports_facet' is NA, the facet is ignored.}
}
\value{
A character string representing the sanitized filename prefix.
}
\description{
Constructs a filename prefix based on provided parameters and an optional facet.
The prefix includes the platform, project title, and current time, with optional
inclusion of a facet or group filter. Spaces are replaced with underscores and
the resulting string is sanitized to ensure it is a valid filesystem path.
}
