% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn_deseq_model.R
\name{learn_deseq_model}
\alias{learn_deseq_model}
\title{Learn a DESeq2 Model from Count Data}
\usage{
learn_deseq_model(sd, metadata, design, params)
}
\arguments{
\item{sd}{A data frame of count data.}

\item{metadata}{A data frame with the same number of rows as there are columns in sd, containing the experimental metadata.}

\item{design}{Character vector specifying design factor (a column in the metadata representing experimental groups).}

\item{params}{The list of configuration parameters for the experiment.}
}
\value{
A DESeq2DataSet object after running the DESeq pipeline.
}
\description{
This function creates a DESeqDataSet and runs the DESeq2 pipeline to learn a model based on count data.
}
