% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_biospyder_new.R
\name{load_biospyder_new}
\alias{load_biospyder_new}
\title{Load BioSpyder Data}
\usage{
load_biospyder_new(biospyder_dbs, temposeq_manifest)
}
\arguments{
\item{biospyder_dbs}{Path to the BioSpyder databases.}

\item{temposeq_manifest}{Name of the TempO-Seq assay manifest file.}
}
\value{
A list containing the BioSpyder data frame and related information.
}
\description{
This function loads the BioSpyder assay data from a specified file, typically a CSV format.
}
