% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_count_data.R
\name{load_count_data}
\alias{load_count_data}
\title{Load Count Data from a File}
\usage{
load_count_data(count_data_file, sampledata_sep)
}
\arguments{
\item{count_data_file}{A string giving the path to the count data file.}

\item{sampledata_sep}{The character string to be used for separating columns in the file. Default is "\\t".}
}
\value{
A data frame with genes as rows and samples as columns.
}
\description{
Reads count data from a specified file, typically a delimited text file.
}
