% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_species.R
\name{load_species}
\alias{load_species}
\title{Load Species Specific Data}
\usage{
load_species(species, wiki, manifest)
}
\arguments{
\item{species}{Target species for the analysis.}

\item{wiki}{Path or URL to the WikiPathways data.}

\item{manifest}{Path to the manifest file required for TempO-Seq analysis.}
}
\value{
A list containing various objects pertaining to the species data.
}
\description{
This function loads information for a specified organism used in the analysis.
It requires Bioconductor packages corresponding to the species being analyzed.
}
