% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_reports.R
\name{make_data_reports}
\alias{make_data_reports}
\title{Generate Data Explorer Report}
\usage{
make_data_reports(pars, facet)
}
\arguments{
\item{pars}{A list of parameters used for report generation.}

\item{facet}{A character string specifying the facet for which to generate the report.}
}
\value{
Invisible NULL. The function is called for its side effect of rendering an HTML report.
}
\description{
Renders the data explorer report HTML file for a given facet using the specified parameters.
The function updates the \code{display_group_filter} in \code{pars} based on the facet provided.
}
\examples{
make_data_reports(pars = list(generate_data_explorer_report = TRUE), facet = "treatment")
}
