% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_reports.R
\name{make_runningfisher_report}
\alias{make_runningfisher_report}
\title{Generate Running Fisher Report}
\usage{
make_runningfisher_report(pars, paths, input)
}
\arguments{
\item{pars}{A list of parameters used for report generation.}

\item{paths}{A list of paths used throughout the analysis.}
}
\value{
Invisible NULL. The function is called for its side effect of rendering an HTML report.
}
\description{
Renders the Running Fisher report HTML file for the whole project.
Uses biosets as input.
Only possible if write_additional_output is TRUE (otherwise biosets will not exist).
}
