% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_reports.R
\name{make_summary_report}
\alias{make_summary_report}
\title{Generate Summary Report}
\usage{
make_summary_report(pars, paths)
}
\arguments{
\item{pars}{A list of parameters used for report generation.}

\item{paths}{A list of paths used throughout the analysis.}
}
\value{
Invisible NULL. The function is called for its side effect of rendering an HTML report.
}
\description{
Renders the Summary report HTML file for the whole project using the specified parameters.
}
