% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_facets.R
\name{parse_facets}
\alias{parse_facets}
\title{Parse facet names for generating reports}
\usage{
parse_facets(metadata, params, skip_extra)
}
\arguments{
\item{metadata}{The metadata containing information about the groups.}

\item{params}{The parameters specifying the reports facet and exclude groups.}

\item{skip_extra}{A vector indicating extra groups to skip.}
}
\value{
A vector of unique facets based on which multiple reports will be generated.
If the reports facet is not specified, it returns NA indicating a single report for all groups.
}
\description{
This function takes in metadata, parameters, and a flag to skip extra groups,
and returns the facets based on which multiple reports will be generated.
}
