% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data_case1}
\alias{prepare_data_case1}
\title{Prepare Data for Case 1: No DESeq2 Facet, No Report Facet}
\usage{
prepare_data_case1(
  ddsList,
  overallResListAll,
  overallResListDEGs,
  rldList,
  mergedDEGsList,
  exp_metadata,
  contrasts
)
}
\arguments{
\item{ddsList}{List containing DESeq2 datasets.}

\item{overallResListAll}{List containing all results.}

\item{overallResListDEGs}{List containing DEGs results.}

\item{rldList}{List containing regularized log2 transformed data.}

\item{mergedDEGsList}{List containing merged DEGs across all contrasts.}

\item{exp_metadata}{Experimental metadata data frame.}

\item{contrasts}{Contrasts data frame.}
}
\value{
Environment with prepared data for the report.
}
\description{
This will generate a single report for all samples. All samples will have been processed together in DESeq2.
}
