% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data_case2}
\alias{prepare_data_case2}
\title{Prepare Data for Case 2: No Facet, Yes Display Facet}
\usage{
prepare_data_case2(params, ...)
}
\arguments{
\item{params}{List of parameters used for the analysis.}

\item{...}{Other necessary parameters ...}
}
\value{
Environment with prepared data for the report.
}
\description{
The specifics of the preparation logic need to be filled in.
}
