% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data_case4}
\alias{prepare_data_case4}
\title{Prepare Data for Case 4: Invalid Configuration}
\usage{
prepare_data_case4(params)
}
\arguments{
\item{params}{List of parameters used for the analysis.}
}
\value{
None; function will stop execution with an error message.
}
\description{
Stops the execution if the configuration is invalid.
}
