% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_facet_data.R
\name{prepare_data_for_report}
\alias{prepare_data_for_report}
\title{Prepare Data For Report}
\usage{
prepare_data_for_report(paths, params)
}
\arguments{
\item{paths}{A list of paths including the RData directory.}

\item{params}{A list of parameters used for the analysis.}
}
\value{
An environment containing prepared data objects for reporting.
}
\description{
This function selects the appropriate data preparation function based on the
facet and display facet parameters and applies it to load necessary data objects
for the analysis report.
}
