% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data_and_metadata.R
\name{process_data_and_metadata}
\alias{process_data_and_metadata}
\title{Process Count Data and Experimental Metadata}
\usage{
process_data_and_metadata(
  count_data,
  exp_metadata,
  contrasts,
  intgroup,
  design,
  params
)
}
\arguments{
\item{count_data}{A matrix or data frame of raw count data from transcriptomic analysis.}

\item{exp_metadata}{A data frame containing the experimental metadata.}

\item{contrasts}{A data frame of contrasts used for differential expression.}

\item{intgroup}{A character vector with names of columns used as interest groups in the analysis.}

\item{design}{The name of the column used for the design in the experimental metadata.}

\item{params}{A list of parameters defining specific aspects of processing.}
}
\value{
A list containing processed count data, experimental metadata, and contrasts.
}
\description{
This function applies a series of processing steps to count data and experimental metadata,
which includes filtering, formatting, sorting and performing sanity checks.
}
\examples{
processed_data <- process_data_and_metadata(raw_counts, metadata, contrast_matrix, c("drug", "dose"), "condition", params)
}
