% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_report.R
\name{render_report}
\alias{render_report}
\title{Render an R Markdown Report}
\usage{
render_report(report_in, report_out, render_pars)
}
\arguments{
\item{report_in}{The path to the R Markdown source file.}

\item{report_out}{The path where the rendered report should be saved.}

\item{render_pars}{A list of parameters to pass to the R Markdown document.}
}
\value{
Invisible NULL. The function is called for its side effect of rendering a report.
}
\description{
This function renders an R Markdown report to the specified output file,
using the provided parameters. It generates a random temporary directory
for storing intermediate files during the rendering process and cleans up
afterwards.
}
\examples{
# Render a report 'example_report.Rmd' to 'example_report.html'
render_report(report_in = "example_report.Rmd",
              report_out = "example_report.html",
              render_pars = list(param1 = "value1", param2 = "value2"))
}
