% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_params.R
\name{replace_nulls_in_config}
\alias{replace_nulls_in_config}
\title{Replace NULL Values in Configuration List with NAs}
\usage{
replace_nulls_in_config(params)
}
\arguments{
\item{params}{A list of configuration parameters.}
}
\value{
The configuration list with \code{NULL} values replaced by \code{NA}.
}
\description{
This function replaces \code{NULL} values in a configuration list with \code{NA}
to avoid issues with unassigned list elements during the configuration setup.
}
