% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_vars.R
\name{row_vars}
\alias{row_vars}
\title{Calculate Row Variances}
\usage{
row_vars(x, ...)
}
\arguments{
\item{x}{A numeric matrix or data frame with rows representing variables and columns representing observations.}

\item{...}{Additional arguments to be passed to \code{rowSums} and \code{rowMeans}.}
}
\value{
A numeric vector containing the variance for each row.
}
\description{
This function computes the variance for each row in a numeric matrix or data frame,
with support for \code{NA} values. Rows with one or fewer non-\code{NA} values will have their
variance set to \code{NA}.
}
