% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_up_paths.R
\name{set_up_paths_2}
\alias{set_up_paths_2}
\title{Set up Paths for DEG Output and BMD Output}
\usage{
set_up_paths_2(paths, params, facets, display_facets)
}
\arguments{
\item{paths}{List of existing paths.}

\item{params}{List of parameters with directory paths and project settings.}

\item{facets}{Character vector of facets to generate paths for.}

\item{display_facets}{Character vector of display facets.}
}
\value{
Updated list of paths with DEG and BMD output paths.
}
\description{
This function updates a list of paths with directories for DEG lists and BMD outputs.
It creates directories for each facet if they do not exist.
}
