% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_up_paths.R
\name{set_up_paths_3}
\alias{set_up_paths_3}
\title{Set up Pathway Analysis Paths}
\usage{
set_up_paths_3(paths, params, report_facets)
}
\arguments{
\item{paths}{List of existing paths.}

\item{params}{List of parameters with directory paths and project settings.}

\item{report_facets}{Character vector of display facets for pathway analysis.}
}
\value{
Updated list of paths with pathway analysis paths.
}
\description{
This function updates a list of paths with directories for pathway analysis outputs.
It creates directories for each display facet if they do not exist.
}
