% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data_and_metadata.R
\name{sort_contrasts}
\alias{sort_contrasts}
\title{Sort contrasts based on experimental metadata and a sorting column}
\usage{
sort_contrasts(exp_metadata, exp_contrasts, params, contrast_col = 1)
}
\arguments{
\item{exp_metadata}{A data frame containing the experiment metadata.}

\item{exp_contrasts}{A data frame containing contrast pairs to sort.}

\item{design}{The name of the column in \code{exp_metadata} that corresponds to the design variable.}

\item{sortcol}{The name of the column to sort by.}
}
\value{
A data frame of sorted contrast pairs.
}
\description{
Sorts the contrasts based on the order of entries in a specified sorting column
within the experimental metadata.
}
