% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_data.R
\name{subset_data}
\alias{subset_data}
\title{Subset Count Data to Match Samples in Metadata}
\usage{
subset_data(count_data, exp_metadata)
}
\arguments{
\item{count_data}{A data frame (or DESeq2 model object) containing count data.}

\item{exp_metadata}{A data frame containing the experimental metadata to match against.}
}
\value{
A subsetted count data frame with the columns ordered to
correspond to the samples specified in the metadata.
}
\description{
Aligns and subsets count data to correspond with samples present in the provided experimental metadata.
}
