% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_metadata.R
\name{subset_metadata}
\alias{subset_metadata}
\title{Subset Experiment Metadata Based on Facet and Filter Criteria}
\usage{
subset_metadata(exp_metadata, design, contrasts, current_facet, current_filter)
}
\arguments{
\item{exp_metadata}{A data frame containing the experimental metadata.}

\item{design}{The name of the column used for the experimental design grouping.}

\item{contrasts}{A data frame of contrasts used for differential expression analysis.}

\item{current_facet}{A character string indicating the facet column to filter on.}

\item{current_filter}{A vector of values defining the filter criteria for the facet.}
}
\value{
A list containing the subsetted experimental metadata and contrasts (each as a data frame).
}
\description{
This function subsets the experimental metadata based on facet and filter criteria.
It also subsets the contrasts data frame based on the subsets defined by this criteria.
}
