% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_across_facets.R
\name{summarize_across_facets}
\alias{summarize_across_facets}
\title{Process DESeq2 Results}
\usage{
summarize_across_facets(
  overallResListAll,
  overallResListDEGs,
  filtered_table,
  facets,
  params
)
}
\arguments{
\item{facets}{Vector of facets to process.}

\item{params}{List of parameters for annotation.}

\item{resultsListAll}{List of all results.}

\item{resultsListDEGs}{List of DEG results.}
}
\value{
A list with combined results for all facets and DEGs.
}
\description{
Annotates DESeq2 results and constructs a unified dataframe for all results.
}
