% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_additional_output.R
\name{write_additional_output}
\alias{write_additional_output}
\title{Write Additional Output for BMDexpress and Biomarker Analysis}
\usage{
write_additional_output(count_data, exp_metadata, design_to_use, params)
}
\arguments{
\item{count_data}{A matrix or data frame containing count data.}

\item{exp_metadata}{A data frame containing the experimental metadata.}

\item{design_to_use}{The column in \code{exp_metadata} that specifies the design.}

\item{params}{A list containing parameters for the analysis.}
}
\description{
This function processes count data and experimental metadata to write out
files formatted for use with BMDexpress and biomarker analysis. It performs
the CPM normalization, log2 transformation, and aggregation as needed.
}
