% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_biosets.R
\name{write_biosets}
\alias{write_biosets}
\title{Generate biosets}
\usage{
write_biosets(params, facet)
}
\arguments{
\item{params}{A list of parameters from which the ID is constructed.}

\item{facet}{A character string indicating the current facet (subset of data) being processed.}
}
\value{
A character string representing the unique analysis ID.
}
\description{
This function takes DEGs (after R-ODAF filters are applied) and applies additional filters
It writes biosets files and returns a biosets objects
}
