% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_tables.R
\name{write_tables}
\alias{write_tables}
\title{Write DESeq2 Results and Annotations to Files}
\usage{
write_tables(facet, params)
}
\arguments{
\item{facet}{A character string indicating the current facet (subset of data) being processed.}
}
\description{
This function takes the results from \code{DESeq2} analysis, annotates them, and writes out several files
including those necessary for downstream analysis and visualization.
}
