% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_contrasts.R
\name{get_contrasts}
\alias{get_contrasts}
\title{Read Contrasts File}
\usage{
get_contrasts(file_path = NULL, paths)
}
\arguments{
\item{file_path}{A string indicating the file path of the contrasts file to read.}

\item{paths}{A list of paths used throughout the analysis.}
}
\value{
A data frame containing the contrast definitions with columns as character data.
}
\description{
Reads the contrasts from a specified tab-delimited file.
}
