% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_facets.R
\name{get_facets}
\alias{get_facets}
\title{Get Facets from Metadata}
\usage{
get_facets(
  metadata,
  params,
  skip_extra = c("DMSO"),
  single_facet_constant = "single_facet_constant_12345"
)
}
\arguments{
\item{metadata}{A data frame containing the experimental metadata.}

\item{params}{A list of parameters for the analysis.}

\item{skip_extra}{A character string specifying an additional group name to be
excluded from the facets.}

\item{single_facet_constant}{A character string specifying the value to return
if no facets are found.}
}
\value{
A character vector of unique facet values after exclusion.
}
\description{
This function extracts unique values from a specified column in the metadata,
excluding any groups specified by the user and an additional skip value.
}
