% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_reports.R
\name{make_data_reports}
\alias{make_data_reports}
\title{Generate Data Explorer Report}
\usage{
make_data_reports(pars, paths, facet, single_facet_constant = "all")
}
\arguments{
\item{pars}{A list of parameters used for report generation.}

\item{paths}{A list of paths used throughout the analysis.}

\item{facet}{A character string specifying the facet for which to generate the report.}

\item{single_facet_constant}{A character string representing the constant for a single facet. Used internally.}
}
\value{
Invisible NULL. The function is called for its side effect of rendering an HTML report.
}
\description{
Renders the data explorer report HTML file for a given facet using the specified parameters.
The function updates the \code{display_group_filter} in \code{pars} based on the facet provided.
}
