% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_up_paths.R
\name{set_up_paths}
\alias{set_up_paths}
\title{Set up Project Paths}
\usage{
set_up_paths(params)
}
\arguments{
\item{params}{List of parameters with directory paths and project settings.}
}
\value{
A list of paths used throughout the analysis.
}
\description{
This function creates a list of paths that the analysis will use to store inputs and outputs.
It ensures that each directory exists, creating them if they do not.
}
